core: update c-ares and use a submodule
We are now swithing to using upstream directly.

This removes support for the TTL field in certain records. This was an
out-of-tree patch I've come to regret.

Not being able too easily update from upstream does more harm than good,
despite my (at the time) good intentions in adding a feature.